+++ src/Oro/Bundle/EntityBundle/Tests/Unit/DependencyInjection/Stub/FirstEntity.php
@@ -1,7 +1,0 @@
-<?php
-
-namespace Oro\Bundle\EntityBundle\Tests\Unit\DependencyInjection\Stub;
-
-class FirstEntity
-{
-}